requirements.txt
Django
djangorestframework
opencv-python-headless
ultralytics  
fast-alpr[onnx-gpu]       
gunicorn      
